PBSpy: Fix problematic choice of STATIC_ROOT in settings.py
Before this commit, the compilestatic+collectstatic commands
works like two stages.
1. The compiled files were stored in pbspy/static/... and
2. collectstatic copies all pbspy/static files (and others...)

This approach does not fit to Django and leads to several problems and
is now replaced by following rule:
The compile command builds the output directly into the target folder
of collectstatic.

Moreover, one entry of STATICFILES_FINDER was commented out because
it is now obsolete and collides with an other finder.